pyproject.toml
[project]
name = "celine-dashboards"
version = "0.1.0"
description = "CELINE dashboards toolkit — Superset SSO plugin, Superset management CLI, Jupyter JWT authorizer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyjwt>=2.10.1",
]

[project.optional-dependencies]
superset = [
    "cryptography>=44.0.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
]
cli = [
    "attrs>=22.2.0",
    "httpx>=0.27.0",
    "python-dateutil>=2.8.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
]
jupyter = []
all = [
    "celine-dashboards[superset,cli,jupyter]",
]

[project.scripts]
celine-superset = "celine.superset.cli.main:app"

[dependency-groups]
dev = [
    "apache-superset==6.0.0",
    "celery>=5.6.0",
    "flask-appbuilder>=5.0.2",
    "flask-caching>=2.3.1",
    "jupyter>=1.1.1",
    "openapi-python-client>=0.28.3",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "python-semantic-release>=9.21.0",
    "respx>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/celine"]

[tool.hatch.build.targets.sdist]
include = ["src/celine", "README.md", "LICENSE"]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
build_command = """
    uv lock --upgrade-package "$PACKAGE_NAME"
    git add uv.lock
    uv build
"""
branch = "main"
changelog_file = "CHANGELOG.md"
tag_format = "v{version}"
commit_parser = "conventional"
upload_to_release = false
upload_to_pypi = false
commit_version_number = false

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf", "up"]
parse_squash_commits = true
ignore_merge_commits = true
